Testing SQLite Connection and Implementation
First test at connecting the SQL and admin dashboard. Used the manage accounts page first, cause that was the main thing that had been made so far for the SQL database. It connects for this version, but it doesn't properly display user names. Connection through SQL is established though, through maven dependencies.